requirements.txt
datasets
matplotlib
pandas
pillow
torch
torchaudio
torchvision
torchmetrics
scipy
opencv-python
setuptools
seaborn
ultralytics
